test gemma 2 fp16

+++ runs.py
@@ -66,7 +66,7 @@ def probe_healthy_gpus(min_free_gb: float = 4.0, compute_test: bool = True) -> l
 DATASET_PATH = "data/HarmBench.csv"
 N_GOALS = None
 RANDOM_SEED = 42
-PRECISION = "int4"
+PRECISION = "fp16"
 
 CHECKPOINT_PATH = f"results/checkpoint_{PRECISION}.json"
